Disable rotation on rM 2. Will likely not get supported in the foreable future.
The feature would require reverse engineering to
find fitting symbols in xochitl/remarkable-shutdown
binaries. Since it isn't needed by the stock ui, this
might not even be possible at all.

After that, people would need to have reason to add
this to the remarkable2-framebuffer repo, which I
would highly doubt as well, since plato was the only
app to utilize this.

Most likely substitute would be to implement a
software-transform instead (the rm1 used hardware
configuration).